Use more sensible configuration

+++ include/dataset.hpp
@@ -27,14 +27,14 @@ struct dataset {
     std::vector<std::vector<double>> test_images;
     std::vector<uint8_t> test_labels;
 
-    std::vector<etl::dyn_matrix<double, 1>> etl_all_images_1d;
-    std::vector<etl::dyn_matrix<double, 1>> etl_training_images_1d;
-    std::vector<etl::dyn_matrix<double, 1>> etl_test_images_1d;
+    std::vector<etl::dyn_matrix<float, 1>> etl_all_images_1d;
+    std::vector<etl::dyn_matrix<float, 1>> etl_training_images_1d;
+    std::vector<etl::dyn_matrix<float, 1>> etl_test_images_1d;
 
     //All the grids
     std::vector<sudoku_grid> source_grids;
 
-    std::vector<etl::dyn_matrix<double, 1>>& training_images_1d(){
+    std::vector<etl::dyn_matrix<float, 1>>& training_images_1d(){
         if(etl_training_images_1d.empty()){
             etl_training_images_1d.reserve(training_images.size());
             for(auto& image : training_images){
@@ -45,7 +45,7 @@ struct dataset {
         return etl_training_images_1d;
     }
 
-    std::vector<etl::dyn_matrix<double, 1>>& test_images_1d(){
+    std::vector<etl::dyn_matrix<float, 1>>& test_images_1d(){
         if(etl_test_images_1d.empty()){
             etl_test_images_1d.reserve(test_images.size());
             for(auto& image : test_images){
@@ -56,7 +56,7 @@ struct dataset {
         return etl_test_images_1d;
     }
 
-    std::vector<etl::dyn_matrix<double, 1>>& all_images_1d(){
+    std::vector<etl::dyn_matrix<float, 1>>& all_images_1d(){
         if(etl_all_images_1d.empty()){
             etl_all_images_1d.reserve(all_images.size());
             for(auto& image : all_images){
